refr(content): move profile components in /components/profile folder
